Fixes bad high port read on Elasticsearch
We had a bug fixed only in the normal json codec around high ports. This
backfills tests to make sure it doesn't arise again.

Thanks to @nollbit for finding this!

Fixes #1350

+++ zipkin-storage/elasticsearch-http/src/main/java/zipkin/storage/elasticsearch/http/ZipkinAdapters.java
@@ -156,7 +156,7 @@ final class ZipkinAdapters {
             result.ipv6(ipv6);
             break;
           case "port":
-            result.port((short) reader.nextInt());
+            result.port(reader.nextInt());
             break;
           default:
             reader.skipValue();
